+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.14 (2023-08-30)
+## 0.0.14 (2023-09-03)
 
 
 ### Bug Fixes
@@ -11,14 +11,17 @@
 ### Features
 
 * add additional folder ([6a3e80e](https://gitlab.com/paras205/boilerplate-react/commit/6a3e80e0283bb533435461c064b30935cd953ddf))
+* add and member list completed ([934b4ea](https://gitlab.com/paras205/boilerplate-react/commit/934b4eab6da534d90b24e3b7fe2844ec5de668b4))
 * added password component, refactored input component ([5676063](https://gitlab.com/paras205/boilerplate-react/commit/5676063ecf247dd38bd7be2ca29c19d3b9c6d76a))
 * added set password template ([75f9471](https://gitlab.com/paras205/boilerplate-react/commit/75f9471f735d8a6cf734924dcc21e0f56326750a))
 * adjusted column width, added edit and add modal in member tier ([827e90e](https://gitlab.com/paras205/boilerplate-react/commit/827e90ebb34d3c5ee73046ad70f05feac8f87be0))
 * admin loagin integartion ([6a58586](https://gitlab.com/paras205/boilerplate-react/commit/6a5858668898a187d75d9633e45aef0ec1867ac1))
 * custom input ([d323c8f](https://gitlab.com/paras205/boilerplate-react/commit/d323c8f84f00f51287e6c564fa41478402ce5ae9))
+* header navbar ([db616a8](https://gitlab.com/paras205/boilerplate-react/commit/db616a8aa8ca6b88e5900b684437b8ebf9b3f345))
 * login component ([f753932](https://gitlab.com/paras205/boilerplate-react/commit/f753932002845eaab1196b776d992ee5ecac0547))
 * login component ([967e159](https://gitlab.com/paras205/boilerplate-react/commit/967e159a4fe973d5e6302f108f2bd08525fcc615))
 * login component ([9da8c3e](https://gitlab.com/paras205/boilerplate-react/commit/9da8c3ed3142a97a439caf5080ed56106b689458))
+* resolved modal issue, integrated api for update property ([56fc7b3](https://gitlab.com/paras205/boilerplate-react/commit/56fc7b3dce0062608340685fa86dcee7ab0d45da))